serde_felt tests
