+++ src/deser/deser.rs
@@ -462,107 +462,3 @@ impl<'de, 'a> SeqAccess<'de> for DeserSeq<'a, 'de> {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[derive(Deserialize, Debug, PartialEq, Eq)]
-    struct Basic {
-        a: FieldElement,
-        b: FieldElement,
-    }
-
-    #[derive(Deserialize, Debug, PartialEq, Eq)]
-    struct Nested {
-        a: FieldElement,
-        b: Basic,
-        c: FieldElement,
-    }
-
-    #[derive(Deserialize, Debug, PartialEq, Eq)]
-    struct WithSequence {
-        a: Vec<FieldElement>,
-        b: FieldElement,
-    }
-
-    #[derive(Deserialize, Debug, PartialEq, Eq)]
-    struct WithArray {
-        a: [FieldElement; 2],
-        b: FieldElement,
-    }
-
-    #[test]
-    fn test_deser_basic() -> Result<()> {
-        let de: Basic = from_felts(&vec![1u64.into(), 2u64.into()])?;
-        let expected = Basic {
-            a: 1u64.into(),
-            b: 2u64.into(),
-        };
-
-        assert_eq!(de, expected);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_deser_nested() -> Result<()> {
-        let de: Nested = from_felts(&vec![1u64.into(), 11u64.into(), 12u64.into(), 2u64.into()])?;
-        let expected = Nested {
-            a: 1u64.into(),
-            b: Basic {
-                a: 11u64.into(),
-                b: 12u64.into(),
-            },
-            c: 2u64.into(),
-        };
-
-        assert_eq!(de, expected);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_deser_seq() -> Result<()> {
-        let de: WithSequence =
-            from_felts(&vec![2u64.into(), 11u64.into(), 12u64.into(), 2u64.into()])?;
-        let expected = WithSequence {
-            a: vec![11u64.into(), 12u64.into()],
-            b: 2u64.into(),
-        };
-
-        assert_eq!(de, expected);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_deser_seq_with_len() -> Result<()> {
-        let len_override = ("a".to_string(), vec![2]);
-        let de: WithSequence = from_felts_with_lengths(
-            &vec![11u64.into(), 12u64.into(), 2u64.into()],
-            vec![len_override].into_iter().collect(),
-        )?;
-        let expected = WithSequence {
-            a: vec![11u64.into(), 12u64.into()],
-            b: 2u64.into(),
-        };
-
-        assert_eq!(de, expected);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_deser_arr() -> Result<()> {
-        let de: WithArray = from_felts(&vec![11u64.into(), 12u64.into(), 2u64.into()])?;
-        let expected = WithArray {
-            a: [11u64.into(), 12u64.into()],
-            b: 2u64.into(),
-        };
-
-        assert_eq!(de, expected);
-
-        Ok(())
-    }
-}
