+++ src/deser/mod.rs
@@ -2,3 +2,6 @@ pub mod deser;
 pub mod error;
 pub mod montgomery;
 pub mod ser;
+
+#[cfg(test)]
+mod tests;
